refactor(insight): split DataProcessor into focused single-responsibility classes
The 1131-line DataProcessor class handled four distinct concerns:
metrics calculation, LLM session analysis, qualitative insight generation,
and orchestration. This splits it into:

- MetricsCalculator (269 lines): heatmaps, streaks, tool usage, durations
- SessionAnalyzer (388 lines): LLM-based session facet analysis and aggregation
- QualitativeInsightGenerator (407 lines): 8 parallel LLM insight calls
- DataProcessor (144 lines): slim orchestrator delegating to the above

Tests updated to test each class directly (no more private method casts).
All 35 existing tests pass unchanged.
